Fill missing currency dates
Fixes #80
